% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivar_awk.R
\name{bivar.awk}
\alias{bivar.awk}
\title{Statistics from a data frame}
\usage{
bivar.awk(mat)
}
\arguments{
\item{mat}{Dataframe of two columns.}
}
\value{
Report with statistics.
}
\description{
Provides a report of number of elements, mean, standard deviation, minimum and maximum for each of two columns and calculate the correlation and regression for the columns.
}
\examples{
df <- data.frame(a = sample(1:1000, 100, replace=TRUE), b = sample(1:1000, 100, replace=TRUE))
bivar.awk(df)

}
