% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeTF-class.R
\docType{class}
\name{CeTF-class}
\alias{CeTF-class}
\alias{CeTF}
\title{The CeTF Class}
\value{
Returns an CeTF object.
}
\description{
The CeTF class is data storage class that stores all the
results from \code{\link{runAnalysis}} function.
}
\section{Slots}{

\describe{
\item{\code{Data}}{Includes the raw, tpm and norm (see \code{\link{normExp}}) data.}

\item{\code{DE}}{Includes the uniquely differentially expressed genes/TFs and
the statistics for all genes (see \code{\link{expDiff}}).}

\item{\code{Input}}{Includes input matrices for RIF (see \code{\link{RIF}}) and PCIT 
(see \code{\link{PCIT}}) for both conditions in \code{\link{runAnalysis}} 
function analysis.}

\item{\code{Output}}{Includes the matrix output from RIF analysis 
(see \code{\link{RIF}}) and a matrix with PCIT output, and other two matrix 
with raw and significant adjacency (see \code{\link{PCIT}}) for both conditions 
inside of \code{\link{runAnalysis}} function analysis.}

\item{\code{Network}}{Network with Gene-Gene and Gene-TF interactions for both 
conditions (see \code{\link{PCIT}}), main TFs resulted from the complete 
analysis, all the TFs identified in the input data and matrix annotating all 
genes and TFs.}
}}

