% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeTFdemo-data.R
\docType{data}
\name{CeTFdemo}
\alias{CeTFdemo}
\title{CeTFdemo class object example}
\format{
An CeTF class object
}
\usage{
data(CeTFdemo)
}
\description{
A CeTFdemo class object to run the examples in functions.
This object was generated after running the runAnalysis function.
This object is the same generated in vignette for complete analysis. Note that
this example data is reduced and don't have the Data slot.
}
\examples{
data(CeTFdemo)
}
\keyword{datasets}
