% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustCoefPercentage.R
\name{clustCoefPercentage}
\alias{clustCoefPercentage}
\title{Calculate the clustering coefficient as a percentage}
\usage{
clustCoefPercentage(mat)
}
\arguments{
\item{mat}{An adjacency matrix. Calculating the clustering coefficient percentage only makes sense if some connections are zero i.e. no connection.}
}
\value{
A numerical between 0 and 100.
}
\description{
Given an adjacency matrix, calculate the clustering coefficient as a percentage of non-zero adjacencies.
}
\examples{
data('simNorm')
results <- PCIT(simNorm)
clustCoefPercentage(results$adj_sig)

}
