% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsToTPM.R
\name{countsToTPM}
\alias{countsToTPM}
\title{Counts to TPM}
\usage{
countsToTPM(counts)
}
\arguments{
\item{counts}{A matrix or dataframe of counts}
}
\value{
Returns a table with TPM values
}
\description{
Convert the count data to TPM
}
\examples{
data('simCounts')
tpm <- countsToTPM(simCounts)



}
