% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.sig.R
\name{density.sig}
\alias{density.sig}
\title{density plot}
\usage{
\method{density}{sig}(mat1, mat2, type = c(1, 2, 3))
}
\arguments{
\item{mat1}{teste}

\item{mat2}{teste}

\item{type}{teste}
}
\value{
teste.
}
\description{
teste.
}
