% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichdemo-data.R
\docType{data}
\name{enrichdemo}
\alias{enrichdemo}
\title{Enrichment data}
\format{An list}
\usage{
data(enrichdemo)
}
\description{
Enrichemnt result from \code{\link{CeTFdemo}} using the genes 
of condition 1 network.
}
\examples{
data(enrichdemo)
}
\keyword{datasets}
