% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{getAnno}
\alias{getAnno}
\alias{getAnno,pcitrif-method}
\title{Method getAnno.}
\usage{
getAnno(x)

\S4method{getAnno}{pcitrif}(x)
}
\value{
Return the annotation for each gene and TFs
}
\description{
Method getAnno.
}
