% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeTF-class.R
\name{getData}
\alias{getData}
\alias{getData,CeTF-method}
\title{Data accessor for a CeTF class object.}
\usage{
getData(x, type = "raw")

\S4method{getData}{CeTF}(x, type = "raw")
}
\arguments{
\item{x}{\code{\link{CeTF-class}} object}

\item{type}{Type of data: raw, tpm or norm (default: raw)}
}
\value{
Returns the raw, tpm or normalized data.
}
\description{
The Data accessor access the raw, tpm and normalized data from 
\code{\link{runAnalysis}} function analysis.
}
\examples{
# load the CeTF class object resulted from runAnalysis function
data(CeTFdemo)

getData(CeTFdemo)

}
\seealso{
\code{\link{runAnalysis}}.
}
