% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{getKeyTF}
\alias{getKeyTF}
\alias{getKeyTF,pcitrif-method}
\title{Method getKeyTF.}
\usage{
getKeyTF(x)

\S4method{getKeyTF}{pcitrif}(x)
}
\value{
Return the main TFs after PCIT and RIF analysis
}
\description{
Method getKeyTF.
}
