% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{getNet1}
\alias{getNet1}
\alias{getNet1,pcitrif-method}
\title{Method getNet1.}
\usage{
getNet1(x)

\S4method{getNet1}{pcitrif}(x)
}
\value{
Return the network of interactions between gene-TFs for condition1
}
\description{
Method getNet1.
}
