% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfToBed.R
\name{gtfToBed}
\alias{gtfToBed}
\title{Converts GTF to BED}
\usage{
gtfToBed(gtf)
}
\arguments{
\item{gtf}{A GTF as data.frame.}
}
\value{
Returns a data.frame in BED format.
}
\description{
Converts a GTF to BED format.
}
