% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeTF-class.R
\name{InputData}
\alias{InputData}
\alias{InputData,CeTF-method}
\title{Input data accessor for a CeTF class object.}
\usage{
InputData(x, analysis = "rif")

\S4method{InputData}{CeTF}(x, analysis = "rif")
}
\arguments{
\item{x}{\code{\link{CeTF-class}} object}

\item{analysis}{Type of analysis: rif, pcit1, pcit2. The numbers 1 and 2 
correspond to the respective condition (default: rif).}
}
\value{
Returns the Inputs used for RIF and PCIT.
}
\description{
The input accessor access the input matrices used for RIF and 
PCIT analysis to both conditions resulted from \code{\link{runAnalysis}} 
function analysis.
}
\examples{
# load the CeTF class object resulted from runAnalysis function
data(CeTFdemo)

InputData(CeTFdemo)

}
\seealso{
\code{\link{runAnalysis}}.
}
