% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netGOplot.R
\name{netGOplot}
\alias{netGOplot}
\title{getGroupGO network plot}
\usage{
netGOplot(netCond, resultsGO, netGO, anno, label = FALSE)
}
\arguments{
\item{netCond}{Network of a specific condition. Can be found in result of \link[pcitRif]{runAnalysis} (step4 -> network_cond1 or network_cond2)}

\item{resultsGO}{Dataframe with the results of \link[pcitRif]{getGroupGO} (first element of list). This result can be decreased applying filters for the pathways selection.}

\item{netGO}{Dataframe with the results of \link[pcitRif]{getGroupGO} (second element of list).}

\item{anno}{Annotation of gene or TFs. Can be found in result of runAnalysis (step4 -> anno)}

\item{label}{If label is TRUE, shows the names of nodes.}
}
\value{
The network for groupGO for a determined condition.
}
\description{
Generate the plot of groupGO network result.
}
\examples{
library(org.Hs.eg.db)

data("pcitrifExample")

genes <- unique(c(as.character(getNet1(pcitrifExample)[,1]),
                 as.character(getNet1(pcitrifExample)[,2])))

cond1 <- getGroupGO(genes = genes,
                    ont = "BP",
                    keyType = "ENSEMBL",
                    annoPkg = org.Hs.eg.db)

t1 <- head(cond1$results, 12)
t2 <- subset(cond1$netGO, cond1$netGO$gene1 \%in\% as.character(t1[,1]))

netGOplot(netCond = getNet1(pcitrifExample),
          resultsGO = t1,
          netGO = t2,
          anno = getAnno(pcitrifExample),
          label = TRUE)




}
