% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeTF-class.R
\name{NetworkData}
\alias{NetworkData}
\alias{NetworkData,CeTF-method}
\title{Networks data accessor for a CeTF class object.}
\usage{
NetworkData(x, type = "network1")

\S4method{NetworkData}{CeTF}(x, type = "network1")
}
\arguments{
\item{x}{\code{\link{CeTF-class}} object}

\item{type}{Type of data: network1, network2, keytfs, tfs or annotation.The 
numbers 1 and 2 correspond to the respective condition (default: network1).}
}
\value{
Returns the Outputs used for RIF and PCIT.
}
\description{
The networks accessor access the networks, key TFs and 
annotations for each gene and TF resulted from PCIT analysis and
\code{\link{runAnalysis}} function analysis.
}
\examples{
# load the CeTF class object resulted from runAnalysis function
data(CeTFdemo)

NetworkData(CeTFdemo)

}
\seealso{
\code{\link{runAnalysis}}.
}
