% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{networkPlot}
\alias{networkPlot}
\title{Final network plot}
\usage{
networkPlot(netCond, netGO, keyTFs)
}
\arguments{
\item{netCond}{Network of a specific condition. Can be found in result of \link[pcitRif]{runAnalysis} (step4 -> network_cond1 or network_cond2)}

\item{netGO}{Dataframe with the results of \link[pcitRif]{getGroupGO} (second element of list). This result can be decreased applying filters for the pathways selection.}

\item{keyTFs}{TFs identified as importants by \link[pcitRif]{runAnalysis} (step4 -> keytf)}
}
\value{
A network.
}
\description{
Generate the network that integrates the network of a condition with groupGO results.
}
\examples{
library(org.Hs.eg.db)

data("pcitrifExample")

genes <- unique(c(as.character(getNet1(pcitrifExample)[,1]),
                 as.character(getNet1(pcitrifExample)[,2])))

cond1 <- getGroupGO(genes = genes,
                    ont = "BP",
                    keyType = "ENSEMBL",
                    annoPkg = org.Hs.eg.db)

t1 <- head(cond1$results, 12)
#Subsetting the network for the conditions to make available only the 12 nodes subsetted
t2 <- subset(cond1$netGO, cond1$netGO$gene1 \%in\% as.character(t1[,1]))

networkPlot(netCond = getNet1(pcitrifExample),
            netGO = t2,
            keyTFs = getKeyTF(pcitrifExample))



}
