% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{normData}
\alias{normData}
\alias{normData,pcitrif-method}
\title{Method normData.}
\usage{
normData(x)

\S4method{normData}{pcitrif}(x)
}
\value{
Return the normalized data
}
\description{
Method normData.
}
