% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normExp.R
\name{normExp}
\alias{normExp}
\title{Normalization of data}
\usage{
normExp(tab)
}
\arguments{
\item{tab}{A matrix or dataframe of expression data}
}
\value{
Returns a table with noralized values
}
\description{
Normalize the expression data of any type of experiment
by columns, applying log(x + 1)/log(2).
}
\examples{
data('simCounts')
tpm <- countsToTPM(simCounts)
norm <- normExp(tpm)



}
