% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCIT.R
\name{PCIT}
\alias{PCIT}
\title{Partial Correlation and Information Theory (PCIT) analysis}
\usage{
PCIT(input, tolType = "mean")
}
\arguments{
\item{input}{A correlation matrix.}

\item{tolType}{Type of tolerance given the 3 pairwise correlations (mean, min, max, median).}
}
\value{
A list with the significant correlations, raw adjacency matrix and significant adjacency matrix.
}
\description{
The PCIT algorithm (Reverter and Chan, 2008) identify the significant correlation of a given matrix of expression.
}
\examples{
data("simNorm")
results <- PCIT(simNorm)
head(results[[1]], 15)

}
