% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{PCITinput1}
\alias{PCITinput1}
\alias{PCITinput1,pcitrif-method}
\title{Method PCITinput1.}
\usage{
PCITinput1(x)

\S4method{PCITinput1}{pcitrif}(x)
}
\value{
Return the PCIT input for condition1
}
\description{
Method PCITinput1.
}
