% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{PCITinput2}
\alias{PCITinput2}
\alias{PCITinput2,pcitrif-method}
\title{Method PCITinput2.}
\usage{
PCITinput2(x)

\S4method{PCITinput2}{pcitrif}(x)
}
\value{
Return the PCIT input for condition2
}
\description{
Method PCITinput2.
}
