% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{PCITout1}
\alias{PCITout1}
\alias{PCITout1,pcitrif-method}
\title{Method PCITout1.}
\usage{
PCITout1(x)

\S4method{PCITout1}{pcitrif}(x)
}
\value{
Return the PCIT output for condition1
}
\description{
Method PCITout1.
}
