% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{PCITout2}
\alias{PCITout2}
\alias{PCITout2,pcitrif-method}
\title{Method PCITout2.}
\usage{
PCITout2(x)

\S4method{PCITout2}{pcitrif}(x)
}
\value{
Return the PCIT output for condition2
}
\description{
Method PCITout2.
}
