% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\docType{class}
\name{pcitrif-class}
\alias{pcitrif-class}
\alias{pcitrif}
\alias{rawData,pcitrif-method}
\alias{tpmData,pcitrif-method}
\alias{tpmDat,pcitrif-method}
\alias{getSelectedGenes,pcitrif-method}
\alias{normData,pcitrif-method}
\alias{getDE,pcitrif-method}
\alias{getTF,pcitrif-method}
\alias{getRIFinput,pcitrif-method}
\alias{getRIFoutput,pcitrif-method}
\alias{getPCITinput_cond1,pcitrif-method}
\alias{getPCITinput_cond2,pcitrif-method}
\alias{getPCIToutput_cond1,pcitrif-method}
\alias{getPCIToutput_cond2,pcitrif-method}
\alias{getNetworkCond1,pcitrif-method}
\alias{getNetworkCond2,pcitrif-method}
\alias{getRIFkeyTF,pcitrif-method}
\alias{getAnno,pcitrif-method}
\title{The pcitrif Class}
\usage{
rawData(x)

tpmData(x)

getSelectedGenes(x)

normData(x)

getDE(x)

getTF(x)

getRIFinput(x)

getRIFoutput(x)

getPCITinput_cond1(x)

getPCITinput_cond2(x)

getPCIToutput_cond1(x)

getPCIToutput_cond2(x)

getNetworkCond1(x)

getNetworkCond2(x)

getRIFkeyTF(x)

getAnno(x)

\S4method{rawData}{pcitrif}(x)

\S4method{tpmData}{pcitrif}(x)

\S4method{getSelectedGenes}{pcitrif}(x)

\S4method{normData}{pcitrif}(x)

\S4method{getDE}{pcitrif}(x)

\S4method{getTF}{pcitrif}(x)

\S4method{getRIFinput}{pcitrif}(x)

\S4method{getRIFoutput}{pcitrif}(x)

\S4method{getPCITinput_cond1}{pcitrif}(x)

\S4method{getPCITinput_cond2}{pcitrif}(x)

\S4method{getPCIToutput_cond1}{pcitrif}(x)

\S4method{getPCIToutput_cond2}{pcitrif}(x)

\S4method{getNetworkCond1}{pcitrif}(x)

\S4method{getNetworkCond2}{pcitrif}(x)

\S4method{getRIFkeyTF}{pcitrif}(x)

\S4method{getAnno}{pcitrif}(x)
}
\description{
The pcitrif class is data storage class that stores all the results of complete RIF and PCIT analysis.
}
\section{Slots}{

\describe{
\item{\code{step1}}{List of objects related to step 1 (TPM Filter) of analysis.}

\item{\code{step2}}{List of objects related to step 2 (Differential Expression) of analysis.}

\item{\code{step3}}{List of objects related to step 3 (Regulatory Impact Factors analysis) of analysis.}

\item{\code{step4}}{List of objects related to step 4 (Partial Correlation and Information Theory analysis) of analysis.}
}}

