% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{rawData}
\alias{rawData}
\alias{rawData,pcitrif-method}
\title{Accessors for the raw data in 'step1' slot of a pcitrif object.}
\usage{
rawData(x)

\S4method{rawData}{pcitrif}(x)
}
\value{
Return the raw data inputed
}
\description{
Accessors for the raw data in 'step1' slot of a pcitrif object.
}
