% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refGenes-data.R
\docType{data}
\name{refGenes}
\alias{refGenes}
\title{List of reference genes for 5 different organisms to perform enrichment}
\format{
An list.
}
\usage{
data(refGenes)
}
\description{
List with protein codings for 5 organisms that must be used as reference 
genes for functional enrichment. This list was generated using Ensembl GTF. 
The organisms are: Human (\emph{Homo sapiens}), Mouse (\emph{Mus musculus}), 
Zebrafish (\emph{Danio rerio}), Cow (\emph{Bos taurus}) and Rat (\emph{Rattus 
norvegicus}).
}
\examples{
data(refGenes)
}
\references{
\url{https://www.ensembl.org/info/data/ftp/index.html}
}
\keyword{datasets}
