% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIF.R
\name{RIF}
\alias{RIF}
\title{Regulatory Impactor Factors analysis}
\usage{
RIF(input, nta = NULL, ntf = NULL, ncond1 = NULL, ncond2 = NULL)
}
\arguments{
\item{input}{teste.}

\item{nta}{teste.}

\item{ntf}{teste.}

\item{ncond1}{teste.}

\item{ncond2}{teste.}
}
\value{
teste.
}
\description{
teste.
}
