% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIF_input-data.R
\docType{data}
\name{RIF_input}
\alias{RIF_input}
\title{Regulatory Impact Factors (RIF) input}
\format{An dataframe.}
\usage{
data(RIF_input)
}
\description{
Data used to the examples of RIF analysis. This data was generated based
on simulated counts and normalized data.
}
\examples{
data(RIF_input)
}
\keyword{datasets}
