% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{RIFinput}
\alias{RIFinput}
\alias{RIFinput,pcitrif-method}
\title{Method RIFinput.}
\usage{
RIFinput(x)

\S4method{RIFinput}{pcitrif}(x)
}
\value{
Return the RIF input
}
\description{
Method RIFinput.
}
