% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{RIFout}
\alias{RIFout}
\alias{RIFout,pcitrif-method}
\title{Accessors for the RIF analysis output in 'step3'
slot of a pcitrif object.}
\usage{
RIFout(x)

\S4method{RIFout}{pcitrif}(x)
}
\arguments{
\item{x}{\code{\link{pcitrif-class}} object}
}
\value{
Returns the output from RIF analysis.
}
\description{
The step3 slot holds the Regulatory Impact Factors (RIF)
analysis, that includes the input for RIF (\code{\link{RIFinput}})
and \code{output} (\code{\link{RIFout}}).
}
\examples{
# load the pcitrif class object resulted from runAnalysis function
data(pcitrifExample)

RIFout(pcitrifExample)
}
\seealso{
\code{\link{RIF}} and \code{\link{runAnalysis}}.
}
