% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Whole analysis of Regulatory Impact Factors (RIF) and Partial Correlation and Information Theory analysis (PCIT)}
\usage{
runAnalysis(
  mat,
  conditions = NULL,
  lfc = 2.57,
  padj = 0.05,
  TFs = NULL,
  ncond1 = NULL,
  ncond2 = NULL,
  tolType = "mean",
  diffMethod = "Reverter",
  data.type = "expression"
)
}
\arguments{
\item{mat}{Count data where the rows are genes and coluns the samples (conditions).}

\item{conditions}{A vector of character identifying the names of conditions (i.e. c('normal', 'tumoral')).}

\item{lfc}{logFoldChange module threshold to define a gene as differentially expressed.}

\item{padj}{Sifnificance value to define a gene as differentially expressed.}

\item{TFs}{A vector of character with all transcripts factors of specific organism.}

\item{ncond1}{Number of samples that correspond to first condition.}

\item{ncond2}{Number of samples that correspond to second condition.}

\item{tolType}{Type of tolerance given the 3 pairwise correlations (mean, min, max, median).}

\item{diffMethod}{Method to calculate Differential Expressed (DE) genes (e.g. Reverter or DESEq2).}

\item{data.type}{Type of input data. If is expression (FPKM, TPM, etc) or counts.}
}
\value{
Returns a pcitRif object with output variables of each step of analysis. The outputs can be used to generate the networks in Cytoscape.
}
\description{
This function uses RIF (Reverter and Chan, 2010) and PCIT (Reverter and Chan, 2008) to run the whole pipeline analysis.
}
\examples{
data('simCounts')
out <- runAnalysis(mat = simCounts,
conditions=c('cond1', 'cond2'),
lfc = 2.57,
padj = 0.05,
TFs = paste0('TF_', 1:1000),
ncond1 = 10,
ncond2= 10,
tolType = 'mean',
diffMethod = 'Reverter')

}
