% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCounts-data.R
\docType{data}
\name{simCounts}
\alias{simCounts}
\title{Simulated counts data}
\format{An dataframe.}
\usage{
data(simCounts)
}
\description{
Simulated counts data created using PROPER package.
This data contains 21,000 genes, 1,000 transcript factors
and 20 samples (divided in two conditions).
}
\examples{
data(simCounts)
}
\keyword{datasets}
