% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simNorm-data.R
\docType{data}
\name{simNorm}
\alias{simNorm}
\title{Simulated normalized data}
\format{An dataframe.}
\usage{
data(simNorm)
}
\description{
Simulated normalized data created using PROPER package.
This data contains 69 genes, and 10 samples
(correspondent to only one condition).
}
\examples{
data(simNorm)
}
\keyword{datasets}
