% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleNetworkPlot.R
\name{singleNetworkPlot}
\alias{singleNetworkPlot}
\title{Network plot for each condition}
\usage{
singleNetworkPlot(x)
}
\arguments{
\item{x}{pcitrif object resulted from \link[pcitRif]{runAnalysis} function.}
}
\value{
The network for both conditions.
}
\description{
Generate the network for the conditions.
}
\examples{
data('simCounts')
out <- runAnalysis(mat = simCounts,
conditions=c('cond1', 'cond2'),
lfc = 2.57,
padj = 0.05,
TFs = paste0('TF_', 1:1000),
ncond1 = 10,
ncond2= 10,
tolType = 'mean',
diffMethod = 'Reverter')

singleNetworkPlot(out)



}
