% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmearPlotDE.R
\name{SmearPlotDE}
\alias{SmearPlotDE}
\title{Smear plot for Differentially Expressed genes}
\usage{
SmearPlotDE(object, diffMethod, lfc = 1.5, padjust = 0.05, conditions)
}
\arguments{
\item{object}{pcitrif class object resulted from \code{\link{runAnalysis}} function.}

\item{diffMethod}{Method used to calculate Differentially Expressed (DE)
genes: 'Reverter' or 'DESeq2' (see \code{\link{expDiff}}).}

\item{lfc}{logFoldChange module threshold to define a gene as differentially expressed (default: 1.5).}

\item{padjust}{Significance value to define a gene as differentially expressed in DESeq2 diffMethod option (default: 0.05).}

\item{conditions}{A vector of characters identifying the names of conditions (i.e. c('normal', 'tumoral')).}
}
\value{
Returns an plot of log2(baseMean) by log2FoldChange or difference of
expression for genes and TF obtained from \code{\link{runAnalysis}} function.
}
\description{
Generate an plot for Differentially Expressed (DE) genes that
shows the relationship between log(baseMean) and Difference of Expression
or log2FoldChange. This plot enables to visualize the distribution of DE genes
and TF in both conditions.
}
\examples{
# load the pcitrif class object resulted from runAnalysis function
data(pcitrifExample)

# performing SmearPlotDE plot
SmearPlotDE(object = pcitrifExample,
            diffMethod = 'Reverter',
            lfc = 1.5,
            conditions = c('untrt', 'trt'))



}
