% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmearPlotTF.R
\name{SmearPlotTF}
\alias{SmearPlotTF}
\title{Smear plot for specifics Transcription Factors}
\usage{
SmearPlotTF(
  object,
  diffMethod,
  lfc = 1.5,
  conditions = NULL,
  TF,
  label = FALSE
)
}
\arguments{
\item{object}{pcitrif class object resulted from \code{\link{runAnalysis}} function.}

\item{diffMethod}{Method used to calculate Differentially Expressed (DE)
genes (see \code{\link{expDiff}}).}

\item{lfc}{logFoldChange module threshold to define a gene as differentially expressed (default: 1.5).}

\item{conditions}{A vector of characters identifying the names of conditions (i.e. c('normal', 'tumoral')).}

\item{TF}{Specify a single TF to be visualized for.}

\item{label}{If label is TRUE, shows the names of single TF and its respectives
targets for both conditions (default: FALSE).}
}
\value{
Returns an plot of log2(baseMean) by log2FoldChange or difference of
expression for a single TF and its targets for both conditions.
}
\description{
Generate an plot for a specific TF. The Targets for this TF
will be found based on \code{\link{getNet1}} and \code{\link{getNet2}} networks
and will be possible to visualize the specific TF with the respective
targets taking a look in the relationship between log(baseMean) and Difference
of Expression or log2FoldChange.
}
\examples{
# load the pcitrif class object resulted from runAnalysis function
data(pcitrifExample)

SmearPlotTF(object = pcitrifExample,
            diffMethod = 'Reverter',
            lfc = 1.5,
            conditions = c('untrt', 'trt'),
            TF = 'ENSG00000185917',
            label = FALSE)



}
