% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFs-data.R
\docType{data}
\name{TFs}
\alias{TFs}
\title{Transcripit Factors data}
\format{An character vector with TFs for Homo sapiens.}
\usage{
data(TFs)
}
\description{
Transcripit Factors data from Kai Wang and Hiroki Nishida, 2015 for Homo sapiens.
}
\examples{
data(TFs)
}
\keyword{datasets}
