% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivar_awk.R
\name{bivar.awk}
\alias{bivar.awk}
\title{Summary statistics from two variables}
\usage{
bivar.awk(x)
}
\arguments{
\item{x}{A dataframe with two columns (variables).}
}
\value{
Returns an summary statistics for two variables.
}
\description{
Read two columns of data values (say X and Y)
and computes summary statistics including N, Mean, SD, Min
and Max for X and Y, as well as the correlation between X
and Y and the regression of Y on X.
}
\examples{
# creating a random dataframe with two columns (variables)
tab <- data.frame(a = sample(1:1000, 100, replace=TRUE),
                  b = sample(1:1000, 100, replace=TRUE))

# running bivar.awk function
bivar.awk(tab)

}
