% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{densityPlot}
\alias{densityPlot}
\title{Density Plot of Raw Correlation Coefficients}
\usage{
densityPlot(mat)
}
\arguments{
\item{mat}{An adjacency matrix.}
}
\value{
The density plot of raw correlation coefficients.
}
\description{
Generate the density plot for an adjacency matrix.
}
\examples{
data('simNorm')
results <- PCIT(simNorm)
densityPlot(results$adj_raw)

}
