% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitySig.R
\name{densitySig}
\alias{densitySig}
\title{Density Plot of raw correlation and significant PCIT}
\usage{
densitySig(mat1, mat2, threshold = 0.5)
}
\arguments{
\item{mat1}{Raw adjacency matrix.}

\item{mat2}{Adjacency matrix resulting from PCIT analysis in which has some zero values.}

\item{threshold}{Threshold of correlation module to plot.}
}
\value{
The density plot of raw correlation with significant PCIT values.
}
\description{
Generate the density plot for adjacency matrices.
}
\examples{
data('simNorm')
results <- PCIT(simNorm)

densitySig(mat1 = results$adj_raw,
mat2 = results$adj_sig,
threshold = 0.5)

}
