% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{getAnno}
\alias{getAnno}
\alias{getAnno,pcitrif-method}
\title{Accessors for the annotation of each gene and TFs in 'step4'
slot of a pcitrif object.}
\usage{
getAnno(x)

\S4method{getAnno}{pcitrif}(x)
}
\arguments{
\item{x}{\code{\link{pcitrif-class}} object}
}
\value{
Returns the annotation for each gene and TFs.
}
\description{
The step4 slot holds the Partial Correlation and Information
Theory (PCIT) analysis, that includes the PCIT input, output and the
network with gene-gene/gene-TFs interaction for condition1 and condition2,
besides the key TFs and the annotation (classification) of genes and TFs.
}
\examples{
# load the pcitrif class object resulted from runAnalysis function
data(pcitrifExample)

getAnno(pcitrifExample)
}
\seealso{
\code{\link{runAnalysis}}.
}
