% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeTF-class.R
\name{getDE}
\alias{getDE}
\alias{getDE,CeTF-method}
\title{Differential Expression accessor for a CeTF class object.}
\usage{
getDE(x, type = "unique")

\S4method{getDE}{CeTF}(x, type = "unique")
}
\arguments{
\item{x}{\code{\link{CeTF-class}} object}

\item{type}{Type of DE matrix: unique and all (default: unique)}
}
\value{
Returns the DE genes with the statistics.
}
\description{
The DE accessor access the differential expression resulted from 
\code{\link{runAnalysis}} function analysis.
}
\examples{
# load the CeTF class object resulted from runAnalysis function
data(CeTFdemo)

getDE(CeTFdemo)

}
\seealso{
\code{\link{runAnalysis}}.
}
