% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{getGenes}
\alias{getGenes}
\alias{getGenes,pcitrif-method}
\title{Method getGenes.}
\usage{
getGenes(x)

\S4method{getGenes}{pcitrif}(x)
}
\value{
Return the filtered genes
}
\description{
Method getGenes.
}
