% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroupGO.R
\name{getGroupGO}
\alias{getGroupGO}
\title{Functional Profile of a gene set at specific GO level}
\usage{
getGroupGO(genes, ont = "BP", keyType = NULL, annoPkg = NULL)
}
\arguments{
\item{genes}{Character vector with the genes to perform the functional profile.}

\item{ont}{One of 'MF', 'BP', and 'CC' subontologies.}

\item{keyType}{key type of input gene.}

\item{annoPkg}{Package of annotation of specific organism (OrgDb).}
}
\value{
A list with results of functional profile of genes and a network with the ontology related with which gene.
}
\description{
Functional Profile of a gene set at specific GO level. Given a vector of genes, this function will return the GO profile at a specific level.
}
\examples{
\dontrun{
results <- getGroupGO(genes,
ont = 'BP',
keyType = 'ENSEMBL',
annoPkg = org.Hs.eg.db)
}



}
