% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroupGO.R
\name{getGroupGO}
\alias{getGroupGO}
\title{Functional Profile of a gene set at specific GO level}
\usage{
getGroupGO(genes, ont = "BP", keyType = NULL, annoPkg = NULL)
}
\arguments{
\item{genes}{Character vector with the genes to perform the functional profile.}

\item{ont}{One of 'MF', 'BP', and 'CC' subontologies.}

\item{keyType}{key type of input gene (i.e. 'ENSEMBL', 'SYMBOL', 'ENTREZID').}

\item{annoPkg}{Package of annotation of specific organism (OrgDb).}
}
\value{
A list with the results of the functional profile of the genes and a
network with the ontologies (column 1) and the corresponding
genes (column 2).
}
\description{
Functional Profile of a gene set at specific GO level.
Given a vector of genes, this function will return the GO profile at
a specific level.
}
\examples{
# load the annotation package
library(org.Hs.eg.db)

# load the pcitrif class object resulted from runAnalysis function
data(pcitrifExample)

# getting the genes in network of condition 1
genes <- unique(c(as.character(getNet1(pcitrifExample)[,1]),
                 as.character(getNet1(pcitrifExample)[,2])))

# performing getGroupGO analysis
cond1 <- getGroupGO(genes = genes,
                    ont = 'BP',
                    keyType = 'ENSEMBL',
                    annoPkg = org.Hs.eg.db)



}
