% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{getNet2}
\alias{getNet2}
\alias{getNet2,pcitrif-method}
\title{Method getNet2.}
\usage{
getNet2(x)

\S4method{getNet2}{pcitrif}(x)
}
\value{
Return the network of interactions between gene-TFs for condition2
}
\description{
Method getNet2.
}
