% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histPlot.R
\name{histPlot}
\alias{histPlot}
\title{Histogram of connectivity distribution}
\usage{
histPlot(mat)
}
\arguments{
\item{mat}{Adjacency matrix resulting from PCIT analysis in which has some zero values.}
}
\value{
The histogram of connectivity distribution.
}
\description{
Generate the histogram for adjacency matrix to show the clustering coefficient distribution.
}
\examples{
data("simNorm")
results <- PCIT(simNorm)
histPlot(results[[3]])



}
