% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netGOplot.R
\name{netGOplot}
\alias{netGOplot}
\title{Plot a network for getGroupGO function}
\usage{
netGOplot(
  netCond,
  resultsGO,
  netGO,
  anno,
  groupBy = "pathways",
  TFs = NULL,
  genes = NULL,
  label = FALSE
)
}
\arguments{
\item{netCond}{Network of a specific condition. Can be found in
result of \code{\link{runAnalysis}} (see \code{\link{getNet1}} and
\code{\link{getNet2}}).}

\item{resultsGO}{Dataframe with the results of \code{\link{getGroupGO}}
(first element of list). This result can be filtered by applying filters
for pathways selection.}

\item{netGO}{Dataframe with the results of \code{\link{getGroupGO}}
(second element of list).}

\item{anno}{Annotation of gene or TFs. Can be found in result of
\code{\link{runAnalysis}} function (see \code{\link{getAnno}}).}

\item{groupBy}{Which variables do you want to group? The options are: 'pathways', 'TFs' and 'genes' (default: 'pathways').}

\item{TFs}{A character with selected TFs.}

\item{genes}{A character with selected genes.}

\item{label}{If label is TRUE, shows the names of nodes (default: FALSE).}
}
\value{
Returns a list with the network for \code{\link{getGroupGO}} output
under a condition and the table used to plot the network as list splitted by
groupBy choice.
}
\description{
Generate the plot of groupGO network result of
\code{\link{getGroupGO}} function.
}
\examples{
# load the annotation package
library(org.Hs.eg.db)

# load the pcitrif class object resulted from runAnalysis function
data(pcitrifExample)

# getting the genes in network of condition 1
genes <- unique(c(as.character(getNet1(pcitrifExample)[,1]),
                 as.character(getNet1(pcitrifExample)[,2])))

# performing getGroupGO analysis
cond1 <- getGroupGO(genes = genes,
                    ont = 'BP',
                    keyType = 'ENSEMBL',
                    annoPkg = org.Hs.eg.db)

# selecting only first 12 pathways
t1 <- head(cond1$results, 12)

# subsetting the network to have only the first 12 pathways
t2 <- subset(cond1$netGO, cond1$netGO$gene1 \%in\% as.character(t1[,1]))

# generate the netGOplot grouping by pathways
pt <- netGOplot(netCond = getNet1(pcitrifExample),
                resultsGO = t1,
                netGO = t2,
                anno = getAnno(pcitrifExample),
                groupBy = 'pathways',
                label = TRUE)
# plotting the network and heading the table
pt$plot
head(pt$tab$`GO:0006807`)



}
