% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{networkPlot}
\alias{networkPlot}
\title{Final network plot}
\usage{
networkPlot(netCond, netGO, keyTFs)
}
\arguments{
\item{netCond}{Network of a specific condition. Can be found in result of \link[pcitRif]{runAnalysis} (step4 -> network_cond1 or network_cond2)}

\item{netGO}{Dataframe with the results of \link[pcitRif]{getGroupGO} (second element of list). This result can be decreased applying filters for the pathways selection.}

\item{keyTFs}{TFs identified as importants by \link[pcitRif]{runAnalysis} (step4 -> keytf)}
}
\value{
A network.
}
\description{
Generate the network that integrates the network of a condition with groupGO results.
}
\examples{
\dontrun{
#See vignette for more details
networkPlot(netCond,
netGO,
keyTFs)
}



}
