% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\name{PCITout1}
\alias{PCITout1}
\alias{PCITout1,pcitrif-method}
\title{Accessors for the PCIT output for condition1 in
'step4' slot of a pcitrif object.}
\usage{
PCITout1(x)

\S4method{PCITout1}{pcitrif}(x)
}
\arguments{
\item{x}{\code{\link{pcitrif-class}} object}
}
\value{
Return the PCIT output for condition1.
}
\description{
The step4 slot holds the Partial Correlation and Information
Theory (PCIT) analysis, that includes the PCIT input, output and the
network with gene-gene/gene-TFs interaction for condition1 and condition2,
besides the key TFs and the annotation (classification) of genes and TFs.
}
\examples{
# load the pcitrif class object resulted from runAnalysis function
data(pcitrifExample)

PCITout1(pcitrifExample)
}
\seealso{
\code{\link{PCIT}} and \code{\link{runAnalysis}}.
}
