% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\docType{class}
\name{pcitrif-class}
\alias{pcitrif-class}
\alias{pcitrif}
\title{The pcitrif Class}
\description{
The pcitrif class is data storage class that stores all the results of complete RIF and PCIT analysis.
}
\section{Slots}{

\describe{
\item{\code{step1}}{List of objects related to step 1 (Data Adjustment) of analysis.}

\item{\code{step2}}{List of objects related to step 2 (Differential Expression) of analysis.}

\item{\code{step3}}{List of objects related to step 3 (Regulatory Impact Factors) of analysis.}

\item{\code{step4}}{List of objects related to step 4 (Partial Correlation and Information Theory) of analysis.}
}}

