% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrif-class.R
\docType{class}
\name{pcitrif-class}
\alias{pcitrif-class}
\alias{pcitrif}
\title{The pcitrif Class}
\value{
A pcitrif object.
}
\description{
The pcitrif class is data storage class that stores all the
results from \code{\link{runAnalysis}} function.
}
\section{Slots}{

\describe{
\item{\code{step1}}{List of objects related to data adjustment.}

\item{\code{step2}}{List of objects related to differential expression analysis.}

\item{\code{step3}}{List of objects related to Regulatory Impact Factors analysis.}

\item{\code{step4}}{List of objects related to Partial Correlation and Information Theory analysis.}
}}

