% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcitrifExample-data.R
\docType{data}
\name{pcitrifExample}
\alias{pcitrifExample}
\title{pcitrif class object example}
\format{An pcitrif class object}
\usage{
data(pcitrifExample)
}
\description{
A pcitrif class object to run the examples in functions.
This object was generated after running the runAnalysis function.
This object is the same generated in vignette for complete analysis.
}
\examples{
data(pcitrifExample)
}
\keyword{datasets}
