% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Partial Correlation and Information Theory analysis}
\usage{
runAnalysis(
  counts,
  conditions = NULL,
  lfc = 2.57,
  TFs = NULL,
  ncond1 = NULL,
  ncond2 = NULL
)
}
\arguments{
\item{counts}{teste.}

\item{conditions}{teste}

\item{lfc}{teste}

\item{TFs}{teste}

\item{ncond1}{teste}

\item{ncond2}{teste}
}
\value{
teste.
}
\description{
teste.
}
