% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFs-data.R
\docType{data}
\name{TFs}
\alias{TFs}
\title{Transcripition Factors data}
\format{An character vector with TFs for \emph{Homo sapiens}.}
\usage{
data(TFs)
}
\description{
Transcripition Factors data from Kai Wang and Hiroki Nishida, 2015
for \emph{Homo sapiens}.
}
\examples{
data(TFs)
}
\references{
\url{https://doi.org/10.1186/s12859-015-0552-x}
}
\keyword{datasets}
