% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.R
\name{tolerance}
\alias{tolerance}
\title{Tolerance}
\usage{
tolerance(a, b, c, type = c("mean", "min", "max", "median"))
}
\arguments{
\item{a}{teste}

\item{b}{teste}

\item{c}{teste}

\item{type}{teste}
}
\value{
teste.
}
\description{
teste.
}
\examples{
teste

}
