% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.R
\name{tolerance}
\alias{tolerance}
\title{Tolerance between 3 pairwise correlations}
\usage{
tolerance(a, b, c, tolType = "mean")
}
\arguments{
\item{a}{Interactor 1.}

\item{b}{Interactor 2.}

\item{c}{Interactor 3.}

\item{tolType}{Type of tolerance given the 3 pairwise correlations (mean, min, max, median).}
}
\value{
Value of tolerance.
}
\description{
Function for calculating the local tolerance given the 3 pairwise correlations between a, b and c.
}
\examples{
data('simNorm')
tolerance(0.5, -0.65, 0.23, tolType = 'mean')
tolerance(0.5, -0.65, 0.23, tolType = 'max')
tolerance(0.5, -0.65, 0.23, tolType = 'min')
tolerance(0.5, -0.65, 0.23, tolType = 'median')

}
