% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.R
\name{tolerance}
\alias{tolerance}
\title{Tolerance level between 3 pairwise correlations}
\usage{
tolerance(a, b, c, tolType = "mean")
}
\arguments{
\item{a}{Correlation value between the genes A and B.}

\item{b}{Correlation value between the genes B and C.}

\item{c}{Correlation value between the genes A and C.}

\item{tolType}{Calculation type for tolerance (mean, min, max, median) (default: 'mean').}
}
\value{
Returns the value of tolerance.
}
\description{
Calculates the local tolerance for every trio of genes.
}
\examples{
tolerance(0.5, -0.65, 0.23, tolType = 'mean')
tolerance(0.5, -0.65, 0.23, tolType = 'max')
tolerance(0.5, -0.65, 0.23, tolType = 'min')
tolerance(0.5, -0.65, 0.23, tolType = 'median')

}
\seealso{
See vignette for more details about the pairwise correlations.
}
